Fix saving of float lists.
(bzr r0.123.117)

+++ settings-backend/gconf.c
@@ -1321,10 +1321,10 @@ writeListValue (CCSSetting *setting,
 	break;
     case TypeFloat:
 	{
-	    float *item;
+	    gdouble *item;
 	    while (list)
 	    {
-		item = malloc (sizeof (float));
+		item = malloc (sizeof (gdouble));
 		if (item)
 		{
 		    *item = list->data->value.asFloat;
